+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The manager now supports migrations to smoothen the on-boarding experience for users with existing set-ups by
   inspecting and cleaning up their game directories.
   - The legacy addonloader can now automatically be removed if it is no longer needed.
+  - Legacy addonloader add-ons can now be renamed to ensure that they are loaded via GW2Load.
 
 ### Changed